backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
typing-extensions==4.8.0
pydantic==2.5.0
email-validator==2.1.0
bcrypt==4.1.1
cryptography>=41.0.7,<46.0.0
psycopg2-binary==2.9.9
typesense==0.18.0
openai==1.2.4